ai-service/requirements.txt
# NGOClaw AI Service Requirements
# Generated from pyproject.toml

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.25.0

# gRPC
grpcio>=1.60.0
grpcio-tools>=1.60.0

# Configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0

# AI SDKs
google-generativeai>=0.3.0
anthropic>=0.8.0
openai>=1.6.0

# Image Processing
pillow>=10.2.0

# Utilities
python-dotenv>=1.0.0
PyYAML>=6.0

